+++ CHANGES.md
@@ -58,6 +58,7 @@ Fixed:
 - Fix implementation of recursive functions (#934).
 - Fix opam install error with some bash-completion configuration (#980).
 - Make `blank()` source unavailable past is expected duration (#668).
+- Fixed implementation details with `cross` operator.
 
 1.4.0 (29-09-2019)
 =====